+++ src/chapter02/assingining_values/AppliyngCasting.java
@@ -7,12 +7,12 @@ public class AppliyngCasting {
 		short bird = (short)1921222; // Stored as 20678
 		int mammal = (int)9f;
 //		long reptile = (long)192301398193810323; // DOES NOT COMPILE
-		long reptile = (long)192301398193810323L; 
-		
+		long reptile = 192301398193810323L;
+
 		System.out.println(fish);
 		System.out.println(bird);
 		System.out.println(mammal);
 		System.out.println(reptile);
-		
+
 	}
 }
