+++ src/chapter02/questions/question06/CandyCounter.java
@@ -8,7 +8,7 @@ public class CandyCounter {
 
 	public static void main(String[] args) {
 		System.out.print(addCandy(1.4, 2.4f) + ", ");
-		System.out.print(addCandy(1.9, (float) 4) + ", ");
-		System.out.print(addCandy((long) (int) (short) 2, (float) 4));
+		System.out.print(addCandy(1.9, 4) + ", ");
+		System.out.print(addCandy((short) 2, 4));
 	}
 }
